requirements.txt
openai>=1.12.0
pinecone-client>=3.0.0
langchain>=0.1.10
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-pinecone>=0.0.3
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0
python-dotenv>=1.0.0
pypdf>=4.0.1
tiktoken>=0.5.2
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
streamlit>=1.31.0
requests>=2.31.0